+++ shellteam.h
@@ -21,7 +21,7 @@ int _atoi(char *c);
 int compstr(char *s1, char *s2);
 char *scrp_dup(char *c);
 int execute_file(char **cmd);
-int shell_len(const char *scrp);
+size_t shell_len(const char *scrp);
 char *_strtok(char *str, char *delim);
 char **mytoken(char *scrp);
 char *pathfinder(char *strings);
